pyproject.toml
[tool.poetry]
name = "llm-sentiment"
version = "0.1.0"
description = ""
authors = ["Artem Pitertcev <artem.pitertsev@gmail.com>"]
readme = "README.md"
packages = [{include = "llm-sentiment"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.109.2"
"uvicorn" = { version = "0.15.0", extras = ["standard"] }
transformers = "4.37.2"
torch = "2.2.0"
pandas = "*"
numpy = "*"
datasets = "*"
accelerate = "*"
mlflow = "^2.10.2"
python-dotenv = "^1.0.1"
setuptools = "^69.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
